requirements.txt
pytest
selenium
webdriver-manager
pytest-html
pytest-xdist
openpyxl
allure-pytest